Prevent confusing simultaneous appearance when row 0 is first
Sequential rows A,B are normally displayed as

AAAAAAAAA
  BBBBBBBBB

And appear simultaneously (lacking previous overlapping rows and
assuming that B isn't more than 5 seconds away from A).

However, if A is forced to line 0 (bottom) for stylistic reasons:

BBBBBBBBB
  AAAAAAAAA

Then having both appear at the same time is confusing. When this
happens, instead force B to appear at least 1.5 seconds later than A,
so that it is evident that A comes first (unless B must start
scrolling earlier than that, of course).